address comments and add a new test

+++ src/main/java/bio/terra/janitor/service/cleanup/flight/GoogleBucketCleanupStep.java
@@ -1,13 +1,15 @@
 package bio.terra.janitor.service.cleanup.flight;
 
 import bio.terra.cloudres.common.ClientConfig;
-import bio.terra.cloudres.google.storage.BlobCow;
 import bio.terra.cloudres.google.storage.BucketCow;
 import bio.terra.generated.model.CloudResourceUid;
 import bio.terra.janitor.db.JanitorDao;
 import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
+import com.google.cloud.storage.BlobId;
 import com.google.cloud.storage.StorageException;
+import java.util.ArrayList;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,10 +29,20 @@ public class GoogleBucketCleanupStep extends ResourceCleanupStep {
       // Delete all Blobs before deleting bucket.
       String bucketName = resourceUid.getGoogleBucketUid().getBucketName();
       BucketCow bucketCow = storageCow.get(bucketName);
-      bucketCow.list().iterateAll().forEach(BlobCow::delete);
+      if (bucketCow == null) {
+        // Bucket doesn't exists.
+        return StepResult.getStepResultSuccess();
+      }
+      List<BlobId> blobIds = new ArrayList<>();
+      bucketCow
+          .list()
+          .iterateAll()
+          .forEach(blobCow -> blobIds.add(blobCow.getBlobInfo().getBlobId()));
+      blobIds.forEach(storageCow::delete);
       bucketCow.delete();
       return StepResult.getStepResultSuccess();
     } catch (StorageException e) {
+      logger.warn("Google StorageException occurs during Bucket Cleanup", e);
       // Catch all exceptions from GOOGLE and consider this retryable error.
       return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, e);
     }
